ArmstrongNumberChecker
This function takes a single number as a parameter and first finds the number of digits (n), then takes each individual digit and raises it to the power of n, adds the total of all 3 powers (stored in the result variable) and finally checks if the result is equal to the number (num) passed in. If yes, returns true. If no, returns false. When the function returns true, this is an Armstrong number.