CodeWars - Who Likes It? (solution)
This is a solution to the CodeWars problem "Who Likes It?". The function is meant to mimic the Facebook "likes" system. An array of names of people who have liked a post is passed in as an argument. Based on the length of the array (how  many likes the post received), the function will return a statement.