CodeWars-Delete_element_appearing_more_than_N_times
This is a solution to the CodeWars Kata called "Delete occurrences of an element if it occurs more than n times":

This function takes in an array of integers and a number (n). It will then loop through the array of integers and delete any that appear more than "n" number of times.