CodeWars-Vasya_Clerk
This is a solution for the CodeWars Kata "Vasya - Clerk":

This function takes in an array of integers (customerCash) as a parameter. The array will contain a combination of 3 integers: 25, 50 and 100 in varying lengths (ex [25, 25, 50, 100, 50, 25]). These integers represent a single customer in the line up to see a movie with either a $25, $50 or $100 dollar bill. A ticket to the movie costs $25. Assuming the movie clerk, Vasya, has no cash on hand, the function will return true if Vasya can properly return exact change to each customer in order of the line or false if not.