CodeWars-Split_strings
This is a solution for the CodeWars Kata "Split Strings":

This function takes in a string (str) as a parameter and splits the string in to pairs of two characters. If the string is odd in length, the second letter in the last pairing will be replaced with an underscore ("_").