CodeWars-Sum_of_digits_digital_roots
This is a solution to the CodeWars Kata "Sum of Digits / Digital Root":

This function takes in a positive integer as a parameter and divides all the single digits in to an array. Each digit is then added together using the built-in reduce function. A while loop is then used to continuously reduce the array down until the sum of all digits is 9 or less.