CodeWars-Take_a_ten_minute_walk
This is a solution to the CodeWars Kata "Take a 10 minute walk":

This function takes in an array of directions (ex. ["n", "s", "e", "w"]) as a parameter. Each direction given counts as walking 1 block. Each block takes 1 minute to walk. If the array does not provide a 10 minute walk (length === 10) and does not return to the exact location, the function returns false. Otherwise if the length of the walk is 10 minutes and returns you to the same spot, the function returns true.