CodeWars-Product_of_consecutive_FIB_nums
This is a solution to the CodeWars Kata "Product of consecutive FIB numbers":

This function takes in an integer as a parameter (prod) and continuously generates fibonacci numbers. If 2 consecutive fibonacci numbers, when multiplied together, are equal to the prod, the function returns both of those numbers and "true". If no 2 consecutive fibonacci numbers multiplied together equal the prod, the function will return the last 2 fibonacci numbers in the array and "false".