CodeWars-the_observed_PIN
This is a solution to the CodeWars Kata "the observed PIN":

This function takes in a string of integers between 1-8 length (o). The value of each integer in the string passed in (o) could be any number above, below, left or right of its placement on the number pad below. The function returns all possible combinations in an array.

1       2       3
4       5       6
7       8       9  
         0