CodeWars-JS-Directions_Reduction
This is a solution in JavaScript for the CodeWars Kata "Directions Reduction":

This function takes in an array of strings that are directions (Ex: ["NORTH", "WEST", "SOUTH", "EAST"]) and removes all unnecessary directions such as moving north and then south right after as this would end up being the same position as the starting point. It loops through the array and removes any pairs of "NORTH"-"SOUTH" or "EAST"-"WEST".