CodeWars-C-Persistent_Bugger
This is a solution in C to the CodeWars Kata "Persistent Bugger":

This function takes in a positive integer, splits the integer in to its individual digits, multiplies all of the digits together and returns the number. If the number is not a single digit, the do/while loop will continue to run until the product of all digits is equal to a single digit integer.